score is now int

+++ src/Move.cpp
@@ -35,7 +35,7 @@ bool Move::isCapture() const { return (moveNum & CAPTURE_FLAG) != 0; }
 // setters
 // Warning! this does not override the already set score of a move. Must be used
 // only if the score of the move is not yet given.
-void Move::setScore(const uint16_t &score) { moveNum |= score << 16; }
+void Move::setScore(const int &score) { moveNum |= score << 16; }
 
 // visualizers
 void Move::print() const {
